добавил авторизацию и аутентификацию, реализовал ограничения для ролей

+++ src/main/java/ru/skypro/homework/service/AdService.java
@@ -2,12 +2,16 @@ package ru.skypro.homework.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import org.webjars.NotFoundException;
 import ru.skypro.homework.dto.AdDTO;
 import ru.skypro.homework.dto.AdsDTO;
 import ru.skypro.homework.dto.CreateOrUpdateAdDTO;
 import ru.skypro.homework.dto.ExtendedAdDTO;
+import ru.skypro.homework.exception.NoAccessException;
+import ru.skypro.homework.exception.NoResultsException;
 import ru.skypro.homework.exception.UserNotAuthorizedException;
 import ru.skypro.homework.model.Ad;
+import ru.skypro.homework.model.Role;
 import ru.skypro.homework.model.User;
 import ru.skypro.homework.service.dao.AdDAO;
 import ru.skypro.homework.service.mapper.AdMapper;
@@ -37,7 +41,7 @@ public class AdService {
         newAd.setTitle(properties.getTitle());
         newAd.setPrice(properties.getPrice());
         newAd.setAuthor(user);
-        newAd.setImage(imageService.uploadAdImage(image));
+        newAd.setImage(imageService.uploadAdImage(properties.getTitle(), image));
 
         adDAO.addAd(newAd);
 
@@ -65,19 +69,50 @@ public class AdService {
         return adMapper.adToExtendedAdDTO(adDAO.getAdById(id));
     }
 
-    public void deleteAd(int id) {
+    public void deleteAd(int id) throws UserNotAuthorizedException, NoAccessException {
+        User user = userService.getAuthorizedUser();
         Ad ad = adDAO.getAdById(id);
-        adDAO.removeAd(ad);
+
+        if (ad.getAuthor().getId() == user.getId() || user.getRole().equals(Role.ADMIN)) {
+            adDAO.removeAd(ad);
+        } else {
+            throw new NoAccessException("No access");
+        }
     }
 
-    public AdDTO updateAd(int id, CreateOrUpdateAdDTO adDTO) {
+    public AdDTO updateAd(int id, CreateOrUpdateAdDTO adDTO) throws UserNotAuthorizedException, NoAccessException {
+        User user = userService.getAuthorizedUser();
         Ad ad = adDAO.getAdById(id);
-        ad.setDescription(adDTO.getDescription());
-        ad.setPrice(adDTO.getPrice());
-        ad.setTitle(ad.getTitle());
 
+        if (ad.getAuthor().getId() == user.getId() || user.getRole().equals(Role.ADMIN)) {
+            ad.setDescription(adDTO.getDescription());
+            ad.setPrice(adDTO.getPrice());
+            ad.setTitle(ad.getTitle());
+            adDAO.updateAd(ad);
+            return adMapper.adToAdDTO(ad);
+        } else {
+            throw new NoAccessException("No access");
+        }
+
+    }
+
+    public byte[] updateAdImage(int adId, MultipartFile image) throws UserNotAuthorizedException, IOException {
+        Ad ad = adDAO.getAdById(adId);
+        ad.setImage(imageService.uploadAdImage(ad.getTitle(), image));
         adDAO.updateAd(ad);
+        return image.getBytes();
+    }
 
-        return adMapper.adToAdDTO(ad);
+    public AdsDTO getAdsWithTitleLike(String query) throws NoResultsException {
+        List<AdDTO> list = adDAO.getAllAds().stream()
+                .filter(ad -> ad.getTitle().contains(query))
+                .map(adMapper::adToAdDTO)
+                .collect(Collectors.toList());
+
+        if (list.isEmpty()) {
+            return new AdsDTO(list.size(), list);
+        } else {
+            throw new NoResultsException("Результатов по запросу нет");
+        }
     }
 }
